# internal/adapters/api/common/openapi.gen.go
// Package common provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package common

// Error defines model for Error.
type Error struct {
	Code      *int    `json:"code,omitempty"`
	Error     string  `json:"error"`
	ErrorType *string `json:"errorType,omitempty"`
	Status    string  `json:"status"`
}

// QueryParamRecords defines model for QueryParamRecords.
type QueryParamRecords = float32

// CreateFakeDataParams defines parameters for CreateFakeData.
type CreateFakeDataParams struct {
	// Number of records to generate in db
	Records QueryParamRecords `form:"records" json:"records"`
}