# internal/adapters/api/transfer/openapi.gen.go
// Package transfer provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package transfer

import (
	"time"
)

// CreateTransferRequest defines model for CreateTransferRequest.
type CreateTransferRequest struct {
	Amount     float32 `json:"amount"`
	ReceiverId int     `json:"receiver_id"`
	SenderId   int     `json:"sender_id"`
}

// Error defines model for Error.
type Error struct {
	Code      *int    `json:"code,omitempty"`
	Error     string  `json:"error"`
	ErrorType *string `json:"errorType,omitempty"`
	Status    string  `json:"status"`
}

// Transfer defines model for Transfer.
type Transfer struct {
	// transfer amount
	Amount float32 `json:"amount"`

	// transfer id
	Id        int        `json:"id"`
	Receiver  Wallet     `json:"receiver"`
	Sender    Wallet     `json:"sender"`
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// Wallet defines model for Wallet.
type Wallet struct {
	// Wallet balance
	Balance float32 `json:"balance"`

	// Wallet id
	Id int `json:"id"`
}

// CreateTransferJSONRequestBody defines body for CreateTransfer for application/json ContentType.
type CreateTransferJSONRequestBody = CreateTransferRequest